working multilingual model without KD

+++ src/MultilingualDataLoader.py
@@ -10,9 +10,8 @@ import io
 from src.utils.model_utils import convert, max_length, PreProcessSentence
 from src.utils.MultilingualUtils import PreProcess
 from src.utils.MultilingualUtils import Padding as padding
-from src.arguments import get_args
 
-languages = ['eng', 'rus', 'ger']
+languages = ['eng', 'ger', 'rus']
 
 def LoadMultlingualDataset(args):
     dataset = {}
@@ -47,29 +46,29 @@ def LoadMultlingualDataset(args):
         src_vocab.fit_on_texts(dataset[lang+'_train_labels'])
         src_vocab.fit_on_texts(dataset[lang+'_train_node1'])
         src_vocab.fit_on_texts(dataset[lang+'_train_node2'])
-
+    exit(0)
+    # Load the vovab is already exists else train one.
     os.makedirs(('vocabs/gat/' + args.lang), exist_ok=True)
-    spm.SentencePieceTrainer.Train('--input=' +target_str +' \
-                                        --model_prefix=vocabs/' + args.model + '/' + args.lang + '/train_tgt \
-                                        --vocab_size=10000 --character_coverage=1.0 --model_type=bpe')
     sp = spm.SentencePieceProcessor()
-    sp.load('vocabs/' + args.model + '/' + args.lang + '/train_tgt.model')
+    if os.path.isfile('vocabs/' + args.model + '/' + args.lang + '/train_tgt.model'):
+        sp.load('vocabs/' + args.model + '/' + args.lang + '/train_tgt.model')
+        print('Vocab file Loaded !\n')
+    else:
+        spm.SentencePieceTrainer.Train('--input=' + target_str + ' \
+                                                    --model_prefix=vocabs/' + args.model + '/' + args.lang + '/train_tgt \
+                                                    --vocab_size='+args.vocab_size+' --character_coverage=1.0 --model_type=bpe')
+        sp.load('vocabs/' + args.model + '/' + args.lang + '/train_tgt.model')
+        with open(('vocabs/gat/' + args.lang + '/' + args.opt + '_src_vocab'), 'wb+') as fp:
+            pickle.dump(src_vocab, fp)
+        print('Vocab file saved !\n')
+
     print('Sentencepiece vocab size {}'.format(sp.get_piece_size()))
-
-    # save the vocab file
-    os.makedirs(('vocabs/gat/' + args.lang), exist_ok=True)
-    with open(('vocabs/gat/' + args.lang + '/' + args.opt + '_src_vocab'), 'wb+') as fp:
-        pickle.dump(src_vocab, fp)
-
-    print('Vocab file saved !\n')
 
     return dataset, src_vocab, sp
 
 def ProcessMultilingualDataset(args):
-    multi_dataset_comps = {}
+    multilingual_dataset = {}
     dataset, src_vocab, tgt_vocab = LoadMultlingualDataset(args)
-    (multi_dataset_comps['nodes'], multi_dataset_comps['labels'],
-     multi_dataset_comps['node1'], multi_dataset_comps['node2']) = [], [], [], []
 
     TRAIN_BUFFER_SIZE = 0
     EVAL_BUFFER_SIZE = 0
@@ -105,6 +104,8 @@ def ProcessMultilingualDataset(args):
                      dataset['ger_train_tgt'].shape[1],
                      dataset['rus_train_tgt'].shape[1],)
 
+    MULTI_BUFFER_SIZE =0
+    BATCH_SIZE = args.batch_size
     for lang in languages:
         dataset[lang + '_train_tgt'] = padding(
             tf.keras.preprocessing.sequence.pad_sequences(dataset[lang + '_train_tgt'],
@@ -113,42 +114,37 @@ def ProcessMultilingualDataset(args):
             tf.keras.preprocessing.sequence.pad_sequences(dataset[lang + '_eval_tgt'],
                                                           padding='post'), MaxSeqSize)
 
-    multilingual_target = tf.concat([dataset[lang + '_train_tgt'] for lang in languages], axis=0)
-    multilingual_nodes = tf.concat([dataset[lang + '_train_nodes'] for lang in languages], axis=0)
-    multilingual_labels = tf.concat([dataset[lang + '_train_labels'] for lang in languages], axis=0)
-    multilingual_node1 = tf.concat([dataset[lang + '_train_node1'] for lang in languages], axis=0)
-    multilingual_node2 = tf.concat([dataset[lang + '_train_node2'] for lang in languages], axis=0)
+        BUFFER_SIZE = len(dataset[lang+'_train_tgt'])
+        MULTI_BUFFER_SIZE += BUFFER_SIZE
+        dataset_size = dataset[lang+'_train_tgt'].shape[0]
 
-    eval_target = tf.concat([dataset[lang + '_eval_tgt'] for lang in languages], axis=0)
-    eval_nodes = tf.concat([dataset[lang + '_eval_nodes'] for lang in languages], axis=0)
-    eval_labels = tf.concat([dataset[lang + '_eval_labels'] for lang in languages], axis=0)
-    eval_node1 = tf.concat([dataset[lang + '_eval_node1'] for lang in languages], axis=0)
-    eval_node2 = tf.concat([dataset[lang + '_eval_node2'] for lang in languages], axis=0)
+        for part in ['train', 'eval']:
+            multilingual_dataset[lang +'_'+ part + '_set'] = tf.data.Dataset.from_tensor_slices(
+                (dataset[lang +'_'+ part + '_nodes'],
+                 dataset[lang +'_'+ part + '_labels'],
+                 dataset[lang +'_'+ part + '_node1'],
+                 dataset[lang +'_'+ part + '_node2'],
+                 dataset[lang +'_'+ part + '_tgt'])).shuffle(BUFFER_SIZE)
+            multilingual_dataset[lang+'_'+part+'_set'] = multilingual_dataset[lang+'_'+part+'_set'].batch(BATCH_SIZE,
+                                                                                                  drop_remainder=True)
+        multilingual_dataset[lang +'_test_set'] = tf.data.Dataset.from_tensor_slices(
+            (dataset[lang +'_test_nodes'],
+             dataset[lang +'_test_labels'],
+             dataset[lang +'_test_node1'],
+             dataset[lang +'_test_node2'])).shuffle(BUFFER_SIZE)
+        multilingual_dataset[lang +'test_set'] = multilingual_dataset[lang +'_test_set'].batch(
+            BATCH_SIZE,
+            drop_remainder=True)
+    final_dataset = {}
+    for opt in ['train', 'test', 'eval']:
+        final_dataset[opt+'_set'] = \
+            multilingual_dataset['eng_'+opt+'_set'].concatenate(
+                multilingual_dataset['ger_'+opt+'_set'].concatenate(
+                    multilingual_dataset['rus_'+opt+'_set']))
 
-    test_nodes = tf.concat([dataset[lang + '_test_nodes'] for lang in languages], axis=0)
-    test_labels = tf.concat([dataset[lang + '_test_labels'] for lang in languages], axis=0)
-    test_node1 = tf.concat([dataset[lang + '_test_node1'] for lang in languages], axis=0)
-    test_node2 = tf.concat([dataset[lang + '_test_node2'] for lang in languages], axis=0)
-    
-    BATCH_SIZE = 32
-    steps_per_epoch = TRAIN_BUFFER_SIZE // BATCH_SIZE
     src_vocab_size = len(src_vocab.word_index) + 1
     tgt_vocab_size = tgt_vocab.get_piece_size()
-    dataset_size = multilingual_target.shape[0]
+    steps_per_epoch = int(MULTI_BUFFER_SIZE // BATCH_SIZE)
 
-    multilingual_dataset = tf.data.Dataset.from_tensor_slices((multilingual_nodes, multilingual_labels,
-                                                               multilingual_node1, multilingual_node2,
-                                                               multilingual_target)).shuffle(TRAIN_BUFFER_SIZE)
-    multilingual_dataset = multilingual_dataset.batch(BATCH_SIZE, drop_remainder=True)
-
-    eval_dataset = tf.data.Dataset.from_tensor_slices((eval_nodes, eval_labels,
-                                                       eval_node1, eval_node2,
-                                                       eval_target)).shuffle(EVAL_BUFFER_SIZE)
-    eval_dataset = eval_dataset.batch(BATCH_SIZE, drop_remainder=True)
-
-    test_set = tf.data.Dataset.from_tensor_slices((test_nodes, test_labels,
-                                                       test_node1, test_node2))
-    test_dataset = test_set.batch(BATCH_SIZE, drop_remainder=True)
-
-    return (multilingual_dataset, eval_dataset, test_dataset, TRAIN_BUFFER_SIZE, EVAL_BUFFER_SIZE, BATCH_SIZE,
-            steps_per_epoch, src_vocab, src_vocab_size, tgt_vocab, tgt_vocab_size, MaxSeqSize, dataset_size)
+    return (final_dataset, src_vocab, src_vocab_size, tgt_vocab,
+            tgt_vocab_size, MULTI_BUFFER_SIZE, steps_per_epoch, MaxSeqSize)
